feat(desktop): add real-time filesystem watcher with @parcel/watcher
Replace polling-based file tree and git status updates with a single
event-driven watcher service. Consolidates StaticPortsWatcher into the
same pattern and decomposes the 500-line filesystem router.

- Add FsWatcher service using @parcel/watcher with 100ms debounce + 2s max batch window
- Add tRPC subscription (filesystem.subscribe) for renderer consumption
- Wire watcher lifecycle to workspace init/delete/close and app startup
- Retire StaticPortsWatcher — ports router now consumes FsWatcher batch events
- Split filesystem router into search.ts, operations.ts, subscription.ts
- Extract shared useFsSubscription hook with optional debounce
- Remove 2.5s polling from ChangesView, use subscription instead
- Add real-time file tree updates in FilesView via subscription
- Auto-invalidate search index cache on filesystem changes
- Delete unused useFileTree.ts and FileTreeNode type
