+++ apps/desktop/electron.vite.config.ts
@@ -84,7 +84,7 @@ export default defineConfig({
 				output: {
 					dir: resolve(devPath, "main"),
 				},
-				external: ["electron", "better-sqlite3", "node-pty"],
+				external: ["electron", "better-sqlite3", "node-pty", "@parcel/watcher"],
 				plugins: [sentryPlugin].filter(Boolean),
 			},
 		},
