+++ apps/desktop/package.json
@@ -36,6 +36,7 @@
 	},
 	"dependencies": {
 		"@anthropic-ai/claude-agent-sdk": "^0.2.19",
+		"@parcel/watcher": "^2.5.0",
 		"@better-auth/stripe": "1.4.18",
 		"@dnd-kit/core": "^6.3.1",
 		"@dnd-kit/sortable": "^10.0.0",
