+++ apps/desktop/scripts/copy-native-modules.ts
@@ -17,7 +17,11 @@ import { cpSync, existsSync, lstatSync, realpathSync, rmSync } from "node:fs";
 import { dirname, join } from "node:path";
 
 // Native modules that must exist for the app to work
-const NATIVE_MODULES = ["better-sqlite3", "node-pty"] as const;
+const NATIVE_MODULES = [
+	"better-sqlite3",
+	"node-pty",
+	"@parcel/watcher",
+] as const;
 
 // Dependencies of native modules that need to be copied (may be hoisted or symlinked)
 const NATIVE_MODULE_DEPS = ["bindings", "file-uri-to-path"] as const;
