+++ apps/desktop/src/main/lib/fs-watcher/index.ts
@@ -1,0 +1,1 @@
+export { fsWatcher } from "./fs-watcher";
