+++ apps/desktop/src/main/lib/static-ports/index.ts
@@ -1,2 +1,1 @@
 export { hasStaticPortsConfig, loadStaticPorts } from "./loader";
-export { staticPortsWatcher } from "./watcher";
