+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/RightSidebar/FilesView/FilesView.tsx
@@ -18,6 +18,7 @@ import { electronTrpc } from "renderer/lib/electron-trpc";
 import { useFileExplorerStore } from "renderer/stores/file-explorer";
 import { useTabsStore } from "renderer/stores/tabs/store";
 import type { DirectoryEntry } from "shared/file-tree-types";
+import { useFsSubscription } from "../hooks/useFsSubscription";
 import { DeleteConfirmDialog } from "./components/DeleteConfirmDialog";
 import { FileSearchResultItem } from "./components/FileSearchResultItem";
 import { FileTreeItem } from "./components/FileTreeItem";
@@ -116,6 +117,11 @@ export function FilesView() {
 		}
 		prevWorktreePathRef.current = worktreePath;
 	}, [worktreePath, tree]);
+
+	useFsSubscription({
+		workspaceId,
+		onData: () => tree.getItemInstance("root")?.invalidateChildrenIds(),
+	});
 
 	const { createFile, createDirectory, rename, deleteItems, isDeleting } =
 		useFileTreeActions({
