+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/RightSidebar/hooks/useFsSubscription/index.ts
@@ -1,0 +1,1 @@
+export { useFsSubscription } from "./useFsSubscription";
