+++ apps/desktop/src/shared/file-tree-types.ts
@@ -1,17 +1,13 @@
-export interface FileTreeNode {
-	id: string;
-	name: string;
-	isDirectory: boolean;
-	path: string;
-	relativePath: string;
-	children?: FileTreeNode[] | null;
-	isLoading?: boolean;
-}
-
 export interface FileSystemChangeEvent {
 	type: "add" | "addDir" | "unlink" | "unlinkDir" | "change";
 	path: string;
 	relativePath: string;
+}
+
+export interface FileSystemBatchEvent {
+	workspaceId: string;
+	events: FileSystemChangeEvent[];
+	timestamp: number;
 }
 
 export interface DirectoryEntry {
